[Dev] up host client

+++ Assets/Dajunctic/Scripts/Lobby/LobbyPlayer.cs
@@ -2,12 +2,12 @@ namespace Dajunctic
 {
     public class LobbyPlayer
     {
-        public ulong ClientId { get; private set; }
+        public int ClientId { get; private set; }
         public string PlayerName { get; private set; }
         public bool IsHost { get; private set; }
         public int PlayerIndex { get; set; }
 
-        public LobbyPlayer(ulong clientId, string playerName, int playerIndex, bool isHost)
+        public LobbyPlayer(int clientId, string playerName, int playerIndex, bool isHost)
         {
             ClientId = clientId;
             PlayerName = playerName;
